--- a/src/absil/il.fsi
+++ b/src/absil/il.fsi
@@ -577,8 +577,8 @@ type ILInstr =
     // Control transfer 
     | I_br    of  ILCodeLabel
     | I_jmp   of ILMethodSpec
-    | I_brcmp of ILComparisonInstr * ILCodeLabel * ILCodeLabel // second label is fall-through 
-    | I_switch    of (ILCodeLabel list * ILCodeLabel) // last label is fallthrough 
+    | I_brcmp of ILComparisonInstr * ILCodeLabel 
+    | I_switch    of ILCodeLabel list 
     | I_ret 
 
      // Method call 
@@ -667,174 +667,38 @@ type ILInstr =
     | EI_ilzero of ILType
     | EI_ldlen_multi      of int32 * int32
 
-/// A list of instructions ending in an unconditionally
-/// branching instruction. A basic block has a label which must be unique
-/// within the method it is located in.  Only the first instruction of
-/// a basic block can be the target of a branch.
-//
-//   Details: The last instruction is always a control flow instruction,
-//   i.e. branch, tailcall, throw etc.
-// 
-//   For example
-//       B1:  ldarg 1
-//            pop
-//            ret
-//
-//   will be one basic block:
-//       ILBasicBlock("B1", [| I_ldarg(1); I_arith(AI_pop); I_ret |])
 
-type ILBasicBlock = 
-    { Label: ILCodeLabel;
-      Instructions: ILInstr[] }
-    member Fallthrough: ILCodeLabel option
+[<RequireQualifiedAccess>]
+type ILExceptionClause = 
+    | Finally of (ILCodeLabel * ILCodeLabel)
+    | Fault  of (ILCodeLabel * ILCodeLabel)
+    | FilterCatch of (ILCodeLabel * ILCodeLabel) * (ILCodeLabel * ILCodeLabel)
+    | TypeCatch of ILType * (ILCodeLabel * ILCodeLabel)
 
+[<RequireQualifiedAccess; NoEquality; NoComparison>]
+type ILExceptionSpec = 
+    { Range: (ILCodeLabel * ILCodeLabel);
+      Clause: ILExceptionClause }
 
 /// Indicates that a particular local variable has a particular source 
-/// language name within a GroupBlock. This does not effect local 
+/// language name within a given set of ranges. This does not effect local 
 /// variable numbering, which is global over the whole method. 
-type ILDebugMapping =
+[<RequireQualifiedAccess; NoEquality; NoComparison>]
+type ILLocalDebugMapping =
     { LocalIndex: int;
       LocalName: string; }
 
-/// ILCode
-/// 
-/// The code for a method is made up of a "code" object.  Each "code"
-/// object gives the contents of the method in a "semi-structured" form, i.e.
-///   1. The structure implicit in the IL exception handling tables
-///      has been made explicit
-///   2. No relative offsets are used in the code: all branches and
-///      switch targets are made explicit as labels.
-///   3. All "fallthroughs" from one basic block to the next have
-///      been made explicit, by adding extra "branch" instructions to
-///      the end of basic blocks which simply fallthrough to another basic
-///      block.
-///
-/// You can convert a straight-line sequence of instructions to structured
-/// code by using buildILCode and 
-/// Most of the interesting code is contained in BasicBlocks. If you're
-/// just interested in getting started with the format then begin
-/// by simply considering methods which do not contain any branch 
-/// instructions, or methods which do not contain any exception handling
-/// constructs.
-///
-/// The above format has the great advantage that you can insert and 
-/// delete new code blocks without needing to fixup relative offsets
-/// or exception tables.  
-///
-/// ILBasicBlock(bblock)
-///   See above
-///
-/// GroupBlock(localDebugInfo, blocks)
-///   A set of blocks, with interior branching between the blocks.  For example
-///       B1:  ldarg 1
-///            br B2
-///
-///       B2:  pop
-///            ret
-///
-///   will be two basic blocks
-///       let b1 = ILBasicBlock("B1", [| I_ldarg(1); I_br("B2") |])
-///       let b2 = ILBasicBlock("B2", [| I_arith(AI_pop); I_ret |])
-///       GroupBlock([], [b1; b2])
-///
-///   A GroupBlock can include a list of debug info records for locally 
-///   scoped local variables.  These indicate that within the given blocks
-///   the given local variables are used for the given Debug info 
-///   will only be recorded for local variables
-///   declared in these nodes, and the local variable will only appear live 
-///   in the debugger for the instructions covered by this node. So if you 
-///   omit or erase these nodes then no debug info will be emitted for local 
-///   variables.  If necessary you can have one outer ScopeBlock which specifies 
-///   the information for all the local variables 
-///  
-///   Not all the destination labels used within a group of blocks need
-///   be satisfied by that group alone.  For example, the interior "try" code
-///   of "try"-"catch" construct may be:
-///       B1:  ldarg 1
-///            br B2
-///
-///       B2:  pop
-///            leave B3
-///
-///   Again there will be two basic blocks grouped together:
-///       let b1 = ILBasicBlock("B1", [| I_ldarg(1); I_br("B2") |])
-///       let b2 = ILBasicBlock("B2", [| I_arith(AI_pop); I_leave("B3") |])
-///       GroupBlock([], [b1; b2])
-///   Here the code must be embedded in a method where "B3" is a label 
-///   somewhere in the method.
-///
-/// RestrictBlock(labels,code) 
-///   This block hides labels, i.e. the given set of labels represent
-///   wiring which is purely internal to the given code block, and may not
-///   be used as the target of a branch by any blocks which this block
-///   is placed alongside.
-///
-///   For example, if a method is made up of:
-///       B1:  ldarg 1
-///            br B2
-///
-///       B2:  ret
-///
-///   then the label "B2" is internal.  The overall code will
-///   be two basic blocks grouped together, surrounded by a RestrictBlock.
-///   The label "B1" is then the only remaining visible entry to the method
-///   and execution will begin at that label.
-///
-///       let b1 = ILBasicBlock("B1", [| I_ldarg(1); I_br("B2") |])
-///       let b2 = ILBasicBlock("B2", [| I_arith(AI_pop); I_leave("B3") |])
-///       let gb1 = GroupBlock([], [b1; b2])
-///       RestrictBlock(["B2"], gb1)
-///
-///   RestrictBlock is necessary to build well-formed code.  
-///
-/// TryBlock(trycode,seh)
-///
-///   A try-catch, try-finally or try-fault block.  
-///   If an exception is raised while executing
-///   an instruction in 'trycode' then the exception handler given by
-///   'seh' is executed.
-///
-/// Well-formedness conditions for code:
-///
-///   Well-formed code includes nodes which explicitly "hide" interior labels.
-///   For example, the code object for a method may have only one entry
-///   label which is not hidden, and this label will be the label where 
-///   execution begins.  
-///
-///   Both filter and catch blocks must have one 
-///   and only one entry.  These entry labels are not visible 
-///   outside the filter and catch blocks. Filter has no 
-///   exits (it always uses endfilter), catch may have exits. 
-///   The "try" block can have multiple entries, i.e. you can branch 
-///   into a try from outside.  They can have multiple exits, each of 
-///   which will be a "leave".
-///
+[<RequireQualifiedAccess; NoEquality; NoComparison>]
+type ILLocalDebugInfo = 
+    { Range: (ILCodeLabel * ILCodeLabel);
+      DebugMappings: ILLocalDebugMapping list }
+
+[<RequireQualifiedAccess; NoEquality; NoComparison>]
 type ILCode = 
-    | ILBasicBlock of ILBasicBlock
-    | GroupBlock of ILDebugMapping list * ILCode list
-    | RestrictBlock of ILCodeLabel list * ILCode
-    | TryBlock of ILCode * ILExceptionBlock
-
-///   The 'seh' specification can have several forms:
-///
-///     FilterCatchBlock
-///       A multi-try-filter-catch block.  Execute the
-///       filters in order to determine which 'catch' block to catch the
-///       exception with. There are two kinds of filters - one for 
-///       filtering exceptions by type and one by an instruction sequence. 
-///       Note that filter blocks can't contain any exception blocks. 
-///
-and ILExceptionBlock = 
-    | FaultBlock of ILCode 
-    | FinallyBlock of ILCode
-    | FilterCatchBlock of (ILFilterBlock * ILCode) list
-
-and ILFilterBlock = 
-    | TypeFilter of ILType
-    | CodeFilter of ILCode
-
-val labelsOfCode: ILCode -> ILCodeLabel list
-val uniqueEntryOfCode: ILCode -> ILCodeLabel
+    { Labels: Dictionary<ILCodeLabel,int> 
+      Instrs:ILInstr[] 
+      Exceptions: ILExceptionSpec list 
+      Locals: ILLocalDebugInfo list }
 
 /// Field Init
 
@@ -855,7 +719,7 @@ type ILFieldInit =
     | Double of double
     | Null
 
-[<RequireQualifiedAccess>]
+[<RequireQualifiedAccess; StructuralEquality; StructuralComparison>]
 type ILNativeVariant = 
     | Empty
     | Null
@@ -947,7 +811,7 @@ type ILNativeType =
 
 
 /// Local variables
-[<NoComparison; NoEquality>]
+[<RequireQualifiedAccess; NoComparison; NoEquality>]
 type ILLocal = 
     { Type: ILType;
       IsPinned: bool;
@@ -957,7 +821,7 @@ type ILLocal =
 type ILLocals = ILList<ILLocal>
 
 /// IL method bodies
-[<NoComparison; NoEquality>]
+[<RequireQualifiedAccess; NoComparison; NoEquality>]
 type ILMethodBody = 
     { IsZeroInit: bool;
       /// strictly speakin should be a uint16 
@@ -1015,6 +879,7 @@ type ILAttributes =
 
 /// Method parameters and return values.
 
+[<RequireQualifiedAccess; NoEquality; NoComparison>]
 type ILParameter = 
     { Name: string option;
       Type: ILType;
@@ -1032,6 +897,7 @@ val typesOfILParamsRaw : ILParameters -> ILTypes
 val typesOfILParamsList : ILParameter list -> ILType list
 
 /// Method return values.
+[<RequireQualifiedAccess; NoEquality; NoComparison>]
 type ILReturn = 
     { Marshal: ILNativeType option;
       Type: ILType; 
@@ -1098,7 +964,7 @@ type PInvokeThrowOnUnmappableChar =
     | Enabled
     | Disabled
 
-[<NoComparison; NoEquality>]
+[<RequireQualifiedAccess; NoComparison; NoEquality>]
 type PInvokeMethod =
     { Where: ILModuleRef;
       Name: string;
@@ -1885,25 +1751,12 @@ val mkILCustomAttribute:
 val mkPermissionSet : ILGlobals -> ILSecurityAction * (ILTypeRef * (string * ILType * ILAttribElem) list) list -> ILPermission
 
 /// Making code.
-val checkILCode:  ILCode -> ILCode
 val generateCodeLabel: unit -> ILCodeLabel
 val formatCodeLabel : ILCodeLabel -> string
 
 /// Make some code that is a straight line sequence of instructions. 
 /// The function will add a "return" if the last instruction is not an exiting instruction.
 val nonBranchingInstrsToCode: ILInstr list -> ILCode 
-
-/// Make some code that is a straight line sequence of instructions, then do 
-/// some control flow.  The first code label is the entry label of the generated code. 
-val mkNonBranchingInstrsThen: ILCodeLabel -> ILInstr list -> ILInstr -> ILCode 
-val mkNonBranchingInstrsThenBr: ILCodeLabel -> ILInstr list -> ILCodeLabel -> ILCode
-
-/// Make a basic block. The final instruction must be control flow.
-val mkNonBranchingInstrs: ILCodeLabel -> ILInstr list -> ILCode
-
-/// Some more primitive helpers.
-val mkBasicBlock: ILBasicBlock -> ILCode
-val mkGroupBlock: ILCodeLabel list * ILCode list -> ILCode
 
 /// Helpers for codegen: scopes for allocating new temporary variables.
 type ILLocalsAllocator =
@@ -2123,20 +1976,6 @@ val rescopeILFieldRef: ILScopeRef -> ILFieldRef -> ILFieldRef
 // The ILCode Builder utility.
 //----------------------------------------------------------------------
 
-[<RequireQualifiedAccess>]
-type ILExceptionClause = 
-    | Finally of (ILCodeLabel * ILCodeLabel)
-    | Fault  of (ILCodeLabel * ILCodeLabel)
-    | FilterCatch of (ILCodeLabel * ILCodeLabel) * (ILCodeLabel * ILCodeLabel)
-    | TypeCatch of ILType * (ILCodeLabel * ILCodeLabel)
-
-type ILExceptionSpec = 
-    { exnRange: (ILCodeLabel * ILCodeLabel);
-      exnClauses: ILExceptionClause list }
-
-type ILLocalSpec = 
-    { locRange: (ILCodeLabel * ILCodeLabel);
-      locInfos: ILDebugMapping list }
 
 /// buildILCode: Build code from a sequence of instructions.
 /// 
@@ -2164,13 +2003,7 @@ type ILLocalSpec =
 /// The input can be badly formed in many ways: exception handlers might
 /// overlap, or scopes of local variables may overlap badly with 
 /// exception handlers.
-val buildILCode:
-    string ->
-    (ILCodeLabel -> int) -> 
-    ILInstr[] -> 
-    ILExceptionSpec list -> 
-    ILLocalSpec list -> 
-    ILCode
+val buildILCode: string -> lab2pc: Dictionary<ILCodeLabel,int> -> instrs:ILInstr[] -> ILExceptionSpec list -> ILLocalDebugInfo list -> ILCode
 
 // -------------------------------------------------------------------- 
 // The instantiation utilities.