This is the last piece of work to remove the "ILX-to-IL" phase from the compiler

Historically, the AbstractIL code representation has long been an overly complicated "basic block" form.  This form was suitable when we needed to do rewriting transformations over the structure that might need to replace instructions with further code.  However, it was expensive and complex to get the generated code into this representation, and then to get it back out again.

Now that the ILX phase has been removed, we can use a much simpler code representation which is just an array of instructions, a dictionary of labels plus information about exceptions and locals.

As part of this, the "branch" instructions now "fall through" to the next instruction, rather than specifying a branch target for the fallthrough case.  This corresponds to the shift to a linear sequence of instructions.

I believe - though do not yet have proof - that this should give better performance for the TAST - to - IL phase of the compiler.  Certainly it reduces the overall code complexity, since this PR removes about 1000 lines from the compiler.  

It's not totally trivial to get the exception and locals information in ilwrite.fs into the form desired, so that part took a while to complete.