@dsyme 2016-06-08
How time flies.... Much of this code was written like 14 years ago! http://research.microsoft.com/pubs/69132/babel01.pdf.  But at least it's still "21st Century Code" :)