@dsyme 2016-06-08
I did some perf tests to see if anything much had been gained.

Looking at compilations of FSharp.Core, and using the ``--times`` flag, and focusing on the last phases of compilation from TAST to binary, we see a reduction from around 3.4sec to 2.9sec (see below for snippets of the output).  That's about 0.5sec out of a total of approx 17sec.  

If repeated on other compilations consistently then that would be about 3%, which is roughly what I stated we expected in earlier PRs.  

BEFORE:
```
ILX0 (1) 3.6
  TIME: 15.1 Delta:  1.5 Mem: 222 G0:   2 G1:  1 G2:  1 [TAST -> ILX]
  TIME: 15.5 Delta:  0.3 Mem: 222 G0:   1 G1:  1 G2:  0 [ILX -> IL (Unions)]
  TIME: 15.5 Delta:  0.0 Mem: 222 G0:   0 G1:  0 G2:  0 [ILX -> IL (Funcs)]
  TIME: 17.3 Delta:  1.8 Mem: 234 G0:   4 G1:  3 G2:  3 [Write .NET Binary]

ILX0 (2) 3.3
  TIME: 14.8 Delta:  1.1 Mem: 247 G0:   1 G1:  1 G2:  0 [TAST -> ILX]
  TIME: 15.1 Delta:  0.3 Mem: 248 G0:   1 G1:  0 G2:  0 [ILX -> IL (Unions)]
  TIME: 15.2 Delta:  0.0 Mem: 248 G0:   0 G1:  0 G2:  0 [ILX -> IL (Funcs)]
  TIME: 17.1 Delta:  1.9 Mem: 239 G0:   4 G1:  4 G2:  3 [Write .NET Binary]

ILX0 (3) 3.4
  TIME: 15.2 Delta:  1.3 Mem: 231 G0:   2 G1:  0 G2:  0 [TAST -> ILX]
  TIME: 15.3 Delta:  0.1 Mem: 231 G0:   0 G1:  0 G2:  0 [ILX -> IL (Unions)]
  TIME: 15.4 Delta:  0.1 Mem: 238 G0:   0 G1:  0 G2:  0 [ILX -> IL (Funcs)]
  TIME: 17.2 Delta:  1.8 Mem: 240 G0:   4 G1:  4 G2:  3 [Write .NET Binary]
```

AFTER:
```
ILX7 (2) 3.1
  TIME: 15.3 Delta:  1.3 Mem: 230 G0:   2 G1:  0 G2:  0 [TAST -> ILX]
  TIME: 17.1 Delta:  1.8 Mem: 244 G0:   4 G1:  4 G2:  3 [Write .NET Binary]

ILX7 (3)  2.9
  TIME: 15.1 Delta:  1.1 Mem: 231 G0:   1 G1:  0 G2:  0 [TAST -> ILX]
  TIME: 16.9 Delta:  1.8 Mem: 237 G0:   4 G1:  3 G2:  3 [Write .NET Binary]

ILX7 (4)  2.8
  TIME: 15.0 Delta:  1.1 Mem: 232 G0:   1 G1:  0 G2:  0 [TAST -> ILX]
  TIME: 16.8 Delta:  1.7 Mem: 238 G0:   4 G1:  3 G2:  3 [Write .NET Binary]
```